+++ src/Adjustments/Changelog.md
@@ -1,5 +1,14 @@
 # Vanilo Adjustments Module Changelog
 
+## 5.x Series
+
+## Unreleased
+##### 2025-XX-XX
+
+- Dropped PHP 8.2 Support
+- Changed the minimum Laravel 10 version to v10.48
+- Added Laravel 12 Support
+
 ## 4.x Series
 
 ## 4.2.0