+++ src/Adjustments/Changelog.md
@@ -7,7 +7,8 @@
 
 - Added the polymorphic `source` relation/fields to the `Adjustment` model
 - Added the `getSource` and `setSource` methods to the `Adjustment` model
-- Changed the `setAdjustable` method to return the itself (was `void` before) 
+- Changed the `setAdjustable` method to return the itself (was `void` before)
+- Changed the minimum Laravel version requirements to v10.48, v11.44 and v12.2 respectively
 
 ## 5.0.0
 ##### 2025-09-03