+++ src/Adjustments/Changelog.md
@@ -2,8 +2,8 @@
 
 ## 5.x Series
 
-## Unreleased
-##### 2025-XX-YY
+## 5.1.0
+##### 2025-12-02
 
 - Added the polymorphic `source` relation/fields to the `Adjustment` model
 - Added the `getSource` and `setSource` methods to the `Adjustment` model